+++ repositories/web-application-stack/images/web-client-tester/app/jest.unit-tests-suite.js
@@ -2,5 +2,5 @@ module.exports = {
     roots: [
         "<rootDir>/tests/Unit/"
     ],
-    testEnvironment: "jsdom"
+    testEnvironment: "node"
 };